Fixed issue 15. Added Versionator class to handle versioning.
